LPS-44643 Reverted "Remove unnecessary abstraction”. Renamed Util class into Impl class
